Discover namespace dynamically
This commit makes the namespace (in which the DaemonSet and its
ServiceAccount run) no longer a "static" resource, and instead
discovered from the operator's running environment. If the operator is
running a cluster, it is discovered from there. If running locally (for
testing purposes) it is discovered from the config.

At the same time we add a bit of code to force the *watch* namespace to
be `""` (i.e. watch all namespaces). This is because the operator needs
to watch the namespace in which the driver runs (so it can reconcile
the DaemonSet) but also any arbitrary namespace in which the user
creates instances of our CR.

Previously it was necessary to set the namespace explicitly when
running locally (`--namespace ''`) and running in a cluster probably
wouldn't have worked at all.