Port in standard `make` stuff
Pending boilerplate standardization of such things, this commit copies
in project.mk and standard.mk from the aws-account-operator and/or
certman-operator repositories, merging/adapting the differences as
appropriate.

We also bring in the standard `coverage` target and utility per
[this SOP](https://github.com/openshift/ops-sop/blob/d35c8edc849a6d22a0645d9655c805fcf62843ec/services/codecov.md),
which prompted a few tweaks to ensure generated files are ignored in the
reports. The `cover` target is renamed `coverhtml` and refactored to use
the results of `coverage`.

The `gofmt` command is tweaked to (again) ignore generated files. And
since it was getting a bit unwieldy for a single line in a `make`
target, it's peeled out into a new helper script. Source files are
modified to make linting happy.