Framework to manage EFS
This is mostly a port of the functionality in
[these scripts](https://gitlab.cee.redhat.com/service/osd-efs/tree/master/scripts)
to manage creation/deletion of EFS file systems and access points.

It's pretty rough right now, not ready for production use, intended for
development, testing, demoing, etc.

You can use it by filling in the structure in pkg/aws/main.go with a
tree of file systems and access points, and then running that `main()`.