Add a helper to clean up
To delete all file systems and access points, you could use a
`desiredState` of `{}`. However, this would still run the unnecessary
step of ensuring the NFS ingress rule exists, which requires discovering
worker nodes and subnets and security groups and...

This is no big deal, really... unless the cluster is in the process of
being torn down, in which case there are no longer worker nodes to
discover, etc. And in this state, the cluster will refuse to finish
uninstalling while file systems exist.

So this commit adds a method, `deleteEverything`, that skips the
discovery and ingress rule creation and just deletes all the mount
targets and file systems (access points come along for the ride).

At the moment, you still have to change configurations (including
selecting `deleteEverything` by editing the main program. A future
commit can make this more CLI-executable friendly.