CLI processing
Now you can:

```shell
$ go build -o manage_efs ./pkg/aws/...
$ ./manage_efs --spec pkg/aws/spec_sample.yaml
... File system and access point construction ensues ...
$ ./manage_efs --delete-all
... File system and access point teardown ensues ...
```

I may want to peel this out into a separate go project. There's no need
for it to be packaged with the operator. However, keeping it in this
repo may make sense, if I'm going to put together a (local) e2e test
that uses it.