Don't try to update PVs
PVs are immutable once created. This *should* mean we can never hit a
code path where we would *need* to `Update` a PV (via `Ensure`). And if
we did, it would bounce anyway because... PVs are immutable once
created.

There's a weird edge case, though: if the shape of a PV created by this
operator changes from one version to the next, the latter will come up
and think the PV needs to be changed to the new shape, and will try to
`Update`. That would bounce and put the operator into a hot loop.

So this commit simply disables `Update`s of PVs by pretending they never
change. We'll still restore one if deleted (which probably also won't
work, for reasons).