Fix references to 0.0.2
Some comments referred to "0.0.2" as the version before we changed the shape of PVs to use the colon-delimited VolumeHandle. But 0.0.2 refers to the [community-operators](https://github.com/operator-framework/community-operators/tree/master/community-operators/aws-efs-operator/0.0.2) (OperatorHub) version, which is just one of many possible version numbers that could refer to the same thing.

Clarify these by instead referring to the commit where the change happened. Let the reader figure out what version that corresponds to in whatever context they care about.