Fix pv[c]Ensurable caching bug
We were neglecting to clear the pvEnsurable and pvcEnsurable caches when
deleting their respective objects. This resulted in behavior where
creating a SharedVolume, deleting it, then recreating one with the same
name but a different spec would result in the *old* spec being used.

Fix this by explicitly clearing the cache entry if we attempt to delete
the resource (even if that deletion fails).

On the test side, completely rework the test case to be fake- rather
than mock-driven.

(The same principle should be applied throughout so we can hopefully get
rid of the mocked client entirely. That'll be the subject of (a)
different PR(s), tracked by [OSD-4098](https://issues.redhat.com/browse/OSD-4098).)

Jira: None